Adds address width generic to AXI4-Lite interface
Introduces an address width generic parameter to the AXI4-Lite interface.

This allows for greater flexibility in address space design, as the address width can now be configured via the generic parameter, rather than being calculated based on the number of registers.

This simplifies address management and improves the overall usability of the AXI4-Lite interface.